--- apps/mobile/components/data-views/collection/EmptyView.tsx
@@ -1,7 +1,0 @@
-import { ScrollView } from 'react-native';
-
-function CollectionEmptyView() {
-	return <ScrollView></ScrollView>;
-}
-
-export default CollectionEmptyView;
