+++ apps/mobile/components/shared/fab/modules/SidebarToggle.tsx
@@ -4,18 +4,15 @@ import { FabMenuItemFactory } from '../fragments/FabMenuItemFactory';
 import { FabMenuItemText } from '../fragments/FabMenuItemIcon';
 import Ionicons from '@expo/vector-icons/Ionicons';
 import { APP_FONT } from '../../../../styles/AppTheme';
-import { useAppDrawerContext } from '../../../../states/useAppDrawer';
 import { useFabController } from '../hooks/useFabController';
 
 const MENU_ITEM_LABEL = 'Sidebar Toggle';
 
 const SidebarToggleModule = memo(({ index }: FabModuleProps) => {
-	const { setOpen } = useAppDrawerContext();
 	const { setIsFabExpanded } = useFabController();
 
 	const onClick = useCallback(() => {
 		try {
-			setOpen(true);
 			setIsFabExpanded(false);
 		} catch (e) {
 			console.log('[WARN]: sidebar context is not hooked to the fab menu');
