+++ apps/mobile/features/timelines/api/useTimeline.ts
@@ -1,290 +1,90 @@
-import { AppTimelineQuery } from './useTimelineController';
-import { useQuery } from '@tanstack/react-query';
+import { useEffect, useState } from 'react';
 import {
-	BlueskyRestClient,
-	DhaagaJsTimelineQueryOptions,
-	KNOWN_SOFTWARE,
-	MisskeyRestClient,
-	PleromaRestClient,
-} from '@dhaaga/bridge';
-import { TimelineFetchMode } from '../../../states/interactors/post-timeline.reducer';
+	useTimelineDispatch,
+	useTimelineState,
+} from '../contexts/PostTimelineCtx';
+import useTimelineQuery from './useTimelineQuery';
 import {
-	useAppAcct,
-	useAppApiClient,
-} from '../../../hooks/utility/global-state-extractors';
-import { AppBskyFeedGetTimeline } from '@atproto/api';
-import { PostMiddleware } from '../../../services/middlewares/post.middleware';
-import { AppPostObject } from '../../../types/app-post.types';
-import { AppResultPageType } from '../../../types/app.types';
+	AppTimelineReducerActionType,
+	TimelineFetchMode,
+} from '../../../states/interactors/post-timeline.reducer';
+import useScrollMoreOnPageEnd from '../../../states/useScrollMoreOnPageEnd';
+import { useAppDb } from '../../../hooks/utility/global-state-extractors';
 
-type TimelineQueryParams = {
-	type: TimelineFetchMode;
-	query?: AppTimelineQuery;
-	opts?: DhaagaJsTimelineQueryOptions;
-	minId?: string;
-	maxId?: string;
-	sessionId?: string;
-};
+function useTimeline(mode: TimelineFetchMode) {
+	const [Refreshing, IsRefreshing] = useState(false);
+	const { db } = useAppDb();
 
-type TimelineFetchResultType = AppResultPageType<AppPostObject>;
+	// state management
+	const State = useTimelineState();
+	const dispatch = useTimelineDispatch();
 
-const DEFAULT_RETURN_VALUE: TimelineFetchResultType = {
-	success: true,
-	items: [],
-	maxId: undefined,
-	minId: undefined,
-};
+	useEffect(() => {
+		if (!db) return;
+		dispatch({
+			type: AppTimelineReducerActionType.INIT,
+			payload: {
+				db,
+			},
+		});
+	}, [db]);
 
-/**
- * Use to fetch paginated data for
- * various timelines containing posts
- *
- * Is consumed by various DataViews
- */
-function useTimeline({
-	type,
-	query,
-	opts,
-	maxId,
-	minId,
-	sessionId,
-}: TimelineQueryParams) {
-	const { client, driver, server } = useAppApiClient();
-	const { acct } = useAppAcct();
+	useEffect(() => {
+		dispatch({
+			type: AppTimelineReducerActionType.RESET_USING_QUERY,
+			payload: {
+				type: mode,
+			},
+		});
+	}, [mode]);
 
-	// to be adjusted based on performance
-	const TIMELINE_STATUS_LIMIT = 10;
-
-	const _id = query?.id;
-	const _query = {
-		// the actual options
-		...opts,
-
-		// injected
-		limit: TIMELINE_STATUS_LIMIT,
-		sinceId: minId,
-		untilId: maxId === null ? undefined : maxId,
-		maxId,
-		minId, // quirks
-		userId: _id,
-	};
-
-	/**
-	 * Generates the cursor based on driver
-	 *
-	 * Only Mastodon and forks are known to embed
-	 * the pagination in headers.
-	 *
-	 * So, either populate maxId in this function
-	 * or use the maxId directly
-	 * @param data
-	 * @param maxId
-	 */
-	function generateMaxId(data: any, maxId?: string | null): string | null {
-		if (maxId) return maxId;
-
-		if (driver === KNOWN_SOFTWARE.BLUESKY) {
-			if (data.posts !== undefined) return data.cursor;
-			const _payload = data as unknown as AppBskyFeedGetTimeline.Response;
-			return _payload.data.cursor;
-		} else {
-			return data[data.length - 1]?.id;
-		}
-	}
-
-	function outputSchemaToResultPage(
-		data: any,
-	): AppResultPageType<AppPostObject> {
-		return {
-			success: true,
-			items: PostMiddleware.deserialize<unknown[]>(data.feed, driver, server),
-			maxId: data.cursor === undefined ? null : data.cursor,
-			minId: null,
-		};
-	}
-
-	function generateFeedBatch(data: any) {
-		let _feed = [];
-		if (driver === KNOWN_SOFTWARE.BLUESKY) {
-			if (data.posts) {
-				// for hashtags
-				_feed = data.posts;
-			} else {
-				const _payload = data as unknown as AppBskyFeedGetTimeline.Response;
-				_feed = _payload.data.feed;
-			}
-		} else {
-			_feed = data.posts ? data.posts : data;
-		}
-		try {
-			return PostMiddleware.deserialize<unknown[]>(_feed, driver, server);
-		} catch (e) {
-			console.log('[ERROR]: failed to convert posts', e);
-			return [];
-		}
-	}
-
-	/**
-	 * aggregated all information that need to
-	 * be calculated to continue the feed
-	 * @param data
-	 * @param maxId
-	 */
-	function createResultBatch(
-		data: any,
-		maxId?: string | null,
-	): TimelineFetchResultType {
-		return {
-			success: true,
-			items: generateFeedBatch(data),
-			maxId: generateMaxId(data, maxId),
-			minId: undefined,
-		};
-	}
-
-	async function api(): Promise<TimelineFetchResultType> {
-		if (client === null) return DEFAULT_RETURN_VALUE;
-		switch (type) {
-			case TimelineFetchMode.IDLE:
-				return DEFAULT_RETURN_VALUE;
-			case TimelineFetchMode.HOME: {
-				const { data, error } = await client.timelines.home(_query);
-				if (error) return DEFAULT_RETURN_VALUE;
-				return createResultBatch(data);
-			}
-			case TimelineFetchMode.LOCAL: {
-				const { data, error } = await client.timelines.public({
-					..._query,
-					local: true, // Pleroma/Akkoma thing
-					withMuted: [KNOWN_SOFTWARE.PLEROMA, KNOWN_SOFTWARE.AKKOMA].includes(
-						driver,
-					)
-						? true
-						: undefined,
-					withRenotes: !opts?.excludeReblogs,
-					withReplies: !opts?.excludeReplies,
-				});
-				if (error) return DEFAULT_RETURN_VALUE;
-				return createResultBatch(data);
-			}
-			case TimelineFetchMode.HASHTAG: {
-				const { data, error } = await client.timelines.hashtag(_id, _query);
-				console.log('all eyes on me', data, error);
-				if (error) return DEFAULT_RETURN_VALUE;
-				return createResultBatch(data);
-			}
-			case TimelineFetchMode.LIST: {
-				const { data, error } = await client.timelines.list(_id, _query);
-				if (error) return DEFAULT_RETURN_VALUE;
-				return createResultBatch(data);
-			}
-			case TimelineFetchMode.USER: {
-				const { data, error } = (await client.accounts.statuses(
-					_id,
-					_query,
-				)) as any;
-				if (error) return DEFAULT_RETURN_VALUE;
-				return createResultBatch(data);
-			}
-			case TimelineFetchMode.SOCIAL: {
-				const { data, error } = await client.timelines.public({
-					..._query,
-					social: true,
-				});
-				if (error) return DEFAULT_RETURN_VALUE;
-				return createResultBatch(data);
-			}
-			case TimelineFetchMode.BUBBLE: {
-				if (driver === KNOWN_SOFTWARE.AKKOMA) {
-					const { data } = await (client as PleromaRestClient).timelines.bubble(
-						_query,
-					);
-					return {
-						success: true,
-						items: PostMiddleware.deserialize<unknown[]>(
-							data as any[],
-							driver,
-							server,
-						),
-						maxId: generateMaxId(data),
-						minId: undefined,
-					};
-				} else if (driver === KNOWN_SOFTWARE.SHARKEY) {
-					const { data } = await (client as MisskeyRestClient).timelines.bubble(
-						_query,
-					);
-					return {
-						success: true,
-						items: PostMiddleware.deserialize<unknown[]>(
-							data as any[],
-							driver,
-							server,
-						),
-						maxId: generateMaxId(data),
-						minId: undefined,
-					};
-				} else {
-					return DEFAULT_RETURN_VALUE;
-				}
-			}
-			case TimelineFetchMode.FEDERATED: {
-				const { data, error } = await client.timelines.public(_query);
-				if (error) return DEFAULT_RETURN_VALUE;
-				return createResultBatch(data);
-			}
-			case TimelineFetchMode.BOOKMARKS: {
-				const { data, error } = await client.accounts.bookmarks(_query);
-				if (error) return DEFAULT_RETURN_VALUE;
-				return createResultBatch(data.data, data?.maxId);
-			}
-			case TimelineFetchMode.FEED: {
-				const { data, error } = await (
-					client as BlueskyRestClient
-				).timelines.feed({
-					limit: 10,
-					cursor: maxId === null ? undefined : maxId,
-					feed: query.id,
-				});
-				if (error) return DEFAULT_RETURN_VALUE;
-				return outputSchemaToResultPage(data);
-			}
-			case TimelineFetchMode.LIKES: {
-				if (driver === KNOWN_SOFTWARE.BLUESKY) {
-					const { data, error } = await (
-						client as BlueskyRestClient
-					).accounts.atProtoLikes(acct.identifier, {
-						limit: 5,
-						cursor: _query.maxId === null ? undefined : _query.maxId,
-					});
-					if (error) return DEFAULT_RETURN_VALUE;
-					return {
-						success: true,
-						items: PostMiddleware.deserialize<unknown[]>(
-							data.feed,
-							driver,
-							server,
-						),
-						maxId: data.cursor === undefined ? null : data.cursor,
-						minId: null,
-					};
-				}
-
-				const { data, error } = await client.accounts.likes(_query);
-				if (error) return DEFAULT_RETURN_VALUE;
-				return outputSchemaToResultPage(data);
-			}
-			default:
-				return DEFAULT_RETURN_VALUE;
-		}
-	}
-
-	// Queries
-	return useQuery<TimelineFetchResultType>({
-		queryKey: [type, _id, _query, maxId, minId, sessionId],
-		queryFn: api,
-		enabled: !!client && type !== TimelineFetchMode.IDLE,
-		initialData: DEFAULT_RETURN_VALUE,
+	const { fetchStatus, data, status, refetch } = useTimelineQuery({
+		type: State.feedType,
+		query: State.query,
+		opts: State.opts,
+		maxId: State.appliedMaxId,
 	});
+
+	useEffect(() => {
+		if (fetchStatus === 'fetching' || status !== 'success') return;
+		dispatch({
+			type: AppTimelineReducerActionType.APPEND_RESULTS,
+			payload: data,
+		});
+	}, [fetchStatus]);
+
+	function loadMore() {
+		dispatch({
+			type: AppTimelineReducerActionType.REQUEST_LOAD_MORE,
+		});
+	}
+
+	/**
+	 * Composite Hook Collection
+	 */
+	const { onScroll, translateY } = useScrollMoreOnPageEnd({
+		itemCount: State.items.length,
+		updateQueryCache: loadMore,
+	});
+
+	function onRefresh() {
+		IsRefreshing(true);
+		dispatch({
+			type: AppTimelineReducerActionType.RESET,
+		});
+		refetch().finally(() => {
+			IsRefreshing(false);
+		});
+	}
+
+	return {
+		refreshing: Refreshing,
+		refresh: onRefresh,
+		data: State.items,
+		onScroll,
+		translateY,
+		fetchStatus,
+	};
 }
 
 export default useTimeline;
